Move team storage functions to Tools
