+++ features/battle/teambuilder.js
@@ -128,105 +128,6 @@ module.exports = {
 	/* Pack Team function - from Pokemon-Showdown-Client */
 
 	packTeam: function (team) {
-		var buf = '';
-		if (!team) return '';
-
-		for (var i = 0; i < team.length; i++) {
-			var set = team[i];
-			if (buf) buf += ']';
-
-			// name
-			buf += (set.name || set.species);
-
-			// species
-			var id = toId(set.species || set.name);
-			buf += '|' + (toId(set.name || set.species) === id ? '' : id);
-
-			// item
-			buf += '|' + toId(set.item);
-
-			// ability
-			var template = set.species || set.name;
-			try {
-				template = require('./../../data/pokedex.js').BattlePokedex[toId(template)];
-			} catch (e) {
-				errlog(e.stack);
-				template = null;
-			}
-			if (!template) return '';
-			var abilities = template.abilities;
-			id = toId(set.ability);
-			if (abilities) {
-				if (abilities['0'] && id === toId(abilities['0'])) {
-					buf += '|';
-				} else if (abilities['1'] && id === toId(abilities['1'])) {
-					buf += '|1';
-				} else if (abilities['H'] && id === toId(abilities['H'])) {
-					buf += '|H';
-				} else {
-					buf += '|' + id;
-				}
-			} else {
-				buf += '|' + id;
-			}
-
-			// moves
-			buf += '|' + set.moves.map(toId).join(',');
-
-			// nature
-			buf += '|' + set.nature;
-
-			// evs
-			var evs = '|';
-			if (set.evs) {
-				evs = '|' + (set.evs['hp'] || '') + ',' + (set.evs['atk'] || '') + ',' + (set.evs['def'] || '') + ',' + (set.evs['spa'] || '') + ',' + (set.evs['spd'] || '') + ',' + (set.evs['spe'] || '');
-			}
-			if (evs === '|,,,,,') {
-				buf += '|';
-			} else {
-				buf += evs;
-			}
-
-			// gender
-			if (set.gender && set.gender !== template.gender) {
-				buf += '|' + set.gender;
-			} else {
-				buf += '|';
-			}
-
-			// ivs
-			var ivs = '|';
-			if (set.ivs) {
-				ivs = '|' + (set.ivs['hp'] === 31 || set.ivs['hp'] === undefined ? '' : set.ivs['hp']) + ',' + (set.ivs['atk'] === 31 || set.ivs['atk'] === undefined ? '' : set.ivs['atk']) + ',' + (set.ivs['def'] === 31 || set.ivs['def'] === undefined ? '' : set.ivs['def']) + ',' + (set.ivs['spa'] === 31 || set.ivs['spa'] === undefined ? '' : set.ivs['spa']) + ',' + (set.ivs['spd'] === 31 || set.ivs['spd'] === undefined ? '' : set.ivs['spd']) + ',' + (set.ivs['spe'] === 31 || set.ivs['spe'] === undefined ? '' : set.ivs['spe']);
-			}
-			if (ivs === '|,,,,,') {
-				buf += '|';
-			} else {
-				buf += ivs;
-			}
-
-			// shiny
-			if (set.shiny) {
-				buf += '|S';
-			} else {
-				buf += '|';
-			}
-
-			// level
-			if (set.level && set.level !== 100) {
-				buf += '|' + set.level;
-			} else {
-				buf += '|';
-			}
-
-			// happiness
-			if (set.happiness !== undefined && set.happiness !== 255) {
-				buf += '|' + set.happiness;
-			} else {
-				buf += '|';
-			}
-		}
-
-		return buf;
+		return Tools.packTeam(team);
 	}
 };
