+++ tools.js
@@ -256,3 +256,765 @@ exports.translateGlobal = function (glob, data, lang) {
 };
 
 loadTranslations();
+
+var BattleStatIDs = exports.BattleStatIDs = {
+	HP: 'hp',
+	hp: 'hp',
+	Atk: 'atk',
+	atk: 'atk',
+	Def: 'def',
+	def: 'def',
+	SpA: 'spa',
+	SAtk: 'spa',
+	SpAtk: 'spa',
+	spa: 'spa',
+	SpD: 'spd',
+	SDef: 'spd',
+	SpDef: 'spd',
+	spd: 'spd',
+	Spe: 'spe',
+	Spd: 'spe',
+	spe: 'spe'
+};
+
+var BattleTypeChart = exports.BattleTypeChart = {
+	"Bug": {
+		damageTaken: {
+			"Bug": 0,
+			"Dark": 0,
+			"Dragon": 0,
+			"Electric": 0,
+			"Fairy": 0,
+			"Fighting": 2,
+			"Fire": 1,
+			"Flying": 1,
+			"Ghost": 0,
+			"Grass": 2,
+			"Ground": 2,
+			"Ice": 0,
+			"Normal": 0,
+			"Poison": 0,
+			"Psychic": 0,
+			"Rock": 1,
+			"Steel": 0,
+			"Water": 0
+		},
+		HPivs: {"atk":30, "def":30, "spd":30}
+	},
+	"Dark": {
+		damageTaken: {
+			"Bug": 1,
+			"Dark": 2,
+			"Dragon": 0,
+			"Electric": 0,
+			"Fairy": 1,
+			"Fighting": 1,
+			"Fire": 0,
+			"Flying": 0,
+			"Ghost": 2,
+			"Grass": 0,
+			"Ground": 0,
+			"Ice": 0,
+			"Normal": 0,
+			"Poison": 0,
+			"Psychic": 3,
+			"Rock": 0,
+			"Steel": 0,
+			"Water": 0
+		},
+		HPivs: {}
+	},
+	"Dragon": {
+		damageTaken: {
+			"Bug": 0,
+			"Dark": 0,
+			"Dragon": 1,
+			"Electric": 2,
+			"Fairy": 1,
+			"Fighting": 0,
+			"Fire": 2,
+			"Flying": 0,
+			"Ghost": 0,
+			"Grass": 2,
+			"Ground": 0,
+			"Ice": 1,
+			"Normal": 0,
+			"Poison": 0,
+			"Psychic": 0,
+			"Rock": 0,
+			"Steel": 0,
+			"Water": 2
+		},
+		HPivs: {"atk":30}
+	},
+	"Electric": {
+		damageTaken: {
+			par: 3,
+			"Bug": 0,
+			"Dark": 0,
+			"Dragon": 0,
+			"Electric": 2,
+			"Fairy": 0,
+			"Fighting": 0,
+			"Fire": 0,
+			"Flying": 2,
+			"Ghost": 0,
+			"Grass": 0,
+			"Ground": 1,
+			"Ice": 0,
+			"Normal": 0,
+			"Poison": 0,
+			"Psychic": 0,
+			"Rock": 0,
+			"Steel": 2,
+			"Water": 0
+		},
+		HPivs: {"spa":30}
+	},
+	"Fairy": {
+		damageTaken: {
+			"Bug": 2,
+			"Dark": 2,
+			"Dragon": 3,
+			"Electric": 0,
+			"Fairy": 0,
+			"Fighting": 2,
+			"Fire": 0,
+			"Flying": 0,
+			"Ghost": 0,
+			"Grass": 0,
+			"Ground": 0,
+			"Ice": 0,
+			"Normal": 0,
+			"Poison": 1,
+			"Psychic": 0,
+			"Rock": 0,
+			"Steel": 1,
+			"Water": 0
+		}
+	},
+	"Fighting": {
+		damageTaken: {
+			"Bug": 2,
+			"Dark": 2,
+			"Dragon": 0,
+			"Electric": 0,
+			"Fairy": 1,
+			"Fighting": 0,
+			"Fire": 0,
+			"Flying": 1,
+			"Ghost": 0,
+			"Grass": 0,
+			"Ground": 0,
+			"Ice": 0,
+			"Normal": 0,
+			"Poison": 0,
+			"Psychic": 1,
+			"Rock": 2,
+			"Steel": 0,
+			"Water": 0
+		},
+		HPivs: {"def":30, "spa":30, "spd":30, "spe":30}
+	},
+	"Fire": {
+		damageTaken: {
+			brn: 3,
+			"Bug": 2,
+			"Dark": 0,
+			"Dragon": 0,
+			"Electric": 0,
+			"Fairy": 2,
+			"Fighting": 0,
+			"Fire": 2,
+			"Flying": 0,
+			"Ghost": 0,
+			"Grass": 2,
+			"Ground": 1,
+			"Ice": 2,
+			"Normal": 0,
+			"Poison": 0,
+			"Psychic": 0,
+			"Rock": 1,
+			"Steel": 2,
+			"Water": 1
+		},
+		HPivs: {"atk":30, "spa":30, "spe":30}
+	},
+	"Flying": {
+		damageTaken: {
+			"Bug": 2,
+			"Dark": 0,
+			"Dragon": 0,
+			"Electric": 1,
+			"Fairy": 0,
+			"Fighting": 2,
+			"Fire": 0,
+			"Flying": 0,
+			"Ghost": 0,
+			"Grass": 2,
+			"Ground": 3,
+			"Ice": 1,
+			"Normal": 0,
+			"Poison": 0,
+			"Psychic": 0,
+			"Rock": 1,
+			"Steel": 0,
+			"Water": 0
+		},
+		HPivs: {"hp":30, "atk":30, "def":30, "spa":30, "spd":30}
+	},
+	"Ghost": {
+		damageTaken: {
+			trapped: 3,
+			"Bug": 2,
+			"Dark": 1,
+			"Dragon": 0,
+			"Electric": 0,
+			"Fairy": 0,
+			"Fighting": 3,
+			"Fire": 0,
+			"Flying": 0,
+			"Ghost": 1,
+			"Grass": 0,
+			"Ground": 0,
+			"Ice": 0,
+			"Normal": 3,
+			"Poison": 2,
+			"Psychic": 0,
+			"Rock": 0,
+			"Steel": 0,
+			"Water": 0
+		},
+		HPivs: {"def":30, "spd":30}
+	},
+	"Grass": {
+		damageTaken: {
+			powder: 3,
+			"Bug": 1,
+			"Dark": 0,
+			"Dragon": 0,
+			"Electric": 2,
+			"Fairy": 0,
+			"Fighting": 0,
+			"Fire": 1,
+			"Flying": 1,
+			"Ghost": 0,
+			"Grass": 2,
+			"Ground": 2,
+			"Ice": 1,
+			"Normal": 0,
+			"Poison": 1,
+			"Psychic": 0,
+			"Rock": 0,
+			"Steel": 0,
+			"Water": 2
+		},
+		HPivs: {"atk":30, "spa":30}
+	},
+	"Ground": {
+		damageTaken: {
+			sandstorm: 3,
+			"Bug": 0,
+			"Dark": 0,
+			"Dragon": 0,
+			"Electric": 3,
+			"Fairy": 0,
+			"Fighting": 0,
+			"Fire": 0,
+			"Flying": 0,
+			"Ghost": 0,
+			"Grass": 1,
+			"Ground": 0,
+			"Ice": 1,
+			"Normal": 0,
+			"Poison": 2,
+			"Psychic": 0,
+			"Rock": 2,
+			"Steel": 0,
+			"Water": 1
+		},
+		HPivs: {"spa":30, "spd":30}
+	},
+	"Ice": {
+		damageTaken: {
+			hail: 3,
+			frz: 3,
+			"Bug": 0,
+			"Dark": 0,
+			"Dragon": 0,
+			"Electric": 0,
+			"Fairy": 0,
+			"Fighting": 1,
+			"Fire": 1,
+			"Flying": 0,
+			"Ghost": 0,
+			"Grass": 0,
+			"Ground": 0,
+			"Ice": 2,
+			"Normal": 0,
+			"Poison": 0,
+			"Psychic": 0,
+			"Rock": 1,
+			"Steel": 1,
+			"Water": 0
+		},
+		HPivs: {"atk":30, "def":30}
+	},
+	"Normal": {
+		damageTaken: {
+			"Bug": 0,
+			"Dark": 0,
+			"Dragon": 0,
+			"Electric": 0,
+			"Fairy": 0,
+			"Fighting": 1,
+			"Fire": 0,
+			"Flying": 0,
+			"Ghost": 3,
+			"Grass": 0,
+			"Ground": 0,
+			"Ice": 0,
+			"Normal": 0,
+			"Poison": 0,
+			"Psychic": 0,
+			"Rock": 0,
+			"Steel": 0,
+			"Water": 0
+		}
+	},
+	"Poison": {
+		damageTaken: {
+			psn: 3,
+			tox: 3,
+			"Bug": 2,
+			"Dark": 0,
+			"Dragon": 0,
+			"Electric": 0,
+			"Fairy": 2,
+			"Fighting": 2,
+			"Fire": 0,
+			"Flying": 0,
+			"Ghost": 0,
+			"Grass": 2,
+			"Ground": 1,
+			"Ice": 0,
+			"Normal": 0,
+			"Poison": 2,
+			"Psychic": 1,
+			"Rock": 0,
+			"Steel": 0,
+			"Water": 0
+		},
+		HPivs: {"def":30, "spa":30, "spd":30}
+	},
+	"Psychic": {
+		damageTaken: {
+			"Bug": 1,
+			"Dark": 1,
+			"Dragon": 0,
+			"Electric": 0,
+			"Fairy": 0,
+			"Fighting": 2,
+			"Fire": 0,
+			"Flying": 0,
+			"Ghost": 1,
+			"Grass": 0,
+			"Ground": 0,
+			"Ice": 0,
+			"Normal": 0,
+			"Poison": 0,
+			"Psychic": 2,
+			"Rock": 0,
+			"Steel": 0,
+			"Water": 0
+		},
+		HPivs: {"atk":30, "spe":30}
+	},
+	"Rock": {
+		damageTaken: {
+			sandstorm: 3,
+			"Bug": 0,
+			"Dark": 0,
+			"Dragon": 0,
+			"Electric": 0,
+			"Fairy": 0,
+			"Fighting": 1,
+			"Fire": 2,
+			"Flying": 2,
+			"Ghost": 0,
+			"Grass": 1,
+			"Ground": 1,
+			"Ice": 0,
+			"Normal": 2,
+			"Poison": 2,
+			"Psychic": 0,
+			"Rock": 0,
+			"Steel": 1,
+			"Water": 1
+		},
+		HPivs: {"def":30, "spd":30, "spe":30}
+	},
+	"Steel": {
+		damageTaken: {
+			psn: 3,
+			tox: 3,
+			sandstorm: 3,
+			"Bug": 2,
+			"Dark": 0,
+			"Dragon": 2,
+			"Electric": 0,
+			"Fairy": 2,
+			"Fighting": 1,
+			"Fire": 1,
+			"Flying": 2,
+			"Ghost": 0,
+			"Grass": 2,
+			"Ground": 1,
+			"Ice": 2,
+			"Normal": 2,
+			"Poison": 3,
+			"Psychic": 2,
+			"Rock": 2,
+			"Steel": 2,
+			"Water": 0
+		},
+		HPivs: {"spd":30}
+	},
+	"Water": {
+		damageTaken: {
+			"Bug": 0,
+			"Dark": 0,
+			"Dragon": 0,
+			"Electric": 1,
+			"Fairy": 0,
+			"Fighting": 0,
+			"Fire": 2,
+			"Flying": 0,
+			"Ghost": 0,
+			"Grass": 1,
+			"Ground": 0,
+			"Ice": 2,
+			"Normal": 0,
+			"Poison": 0,
+			"Psychic": 0,
+			"Rock": 0,
+			"Steel": 2,
+			"Water": 2
+		},
+		HPivs: {"atk":30, "def":30, "spa":30}
+	}
+};
+
+var teamToJSON = exports.teamToJSON = function (text) {
+	text = text.split("\n");
+	var team = [];
+	var curSet = null;
+	for (var i = 0; i < text.length; i++) {
+		var line = text[i].trim();
+		if (line === '' || line === '---') {
+			curSet = null;
+		} else if (!curSet) {
+			curSet = {name: '', species: '', gender: ''};
+			team.push(curSet);
+			var atIndex = line.lastIndexOf(' @ ');
+			if (atIndex !== -1) {
+				curSet.item = line.substr(atIndex + 3);
+				if (toId(curSet.item) === 'noitem') curSet.item = '';
+				line = line.substr(0, atIndex);
+			}
+			if (line.substr(line.length - 4) === ' (M)') {
+				curSet.gender = 'M';
+				line = line.substr(0, line.length - 4);
+			}
+			if (line.substr(line.length - 4) === ' (F)') {
+				curSet.gender = 'F';
+				line = line.substr(0, line.length - 4);
+			}
+			var parenIndex = line.lastIndexOf(' (');
+			if (line.substr(line.length - 1) === ')' && parenIndex !== -1) {
+				line = line.substr(0, line.length - 1);
+				curSet.species = line.substr(parenIndex + 2);
+				line = line.substr(0, parenIndex);
+				curSet.name = line;
+			} else {
+				curSet.species = line;
+				curSet.name = curSet.species;
+			}
+		} else if (line.substr(0, 7) === 'Trait: ') {
+			line = line.substr(7);
+			curSet.ability = line;
+		} else if (line.substr(0, 9) === 'Ability: ') {
+			line = line.substr(9);
+			curSet.ability = line;
+		} else if (line === 'Shiny: Yes') {
+			curSet.shiny = true;
+		} else if (line.substr(0, 7) === 'Level: ') {
+			line = line.substr(7);
+			curSet.level = parseInt(line);
+		} else if (line.substr(0, 11) === 'Happiness: ') {
+			line = line.substr(11);
+			curSet.happiness = parseInt(line);
+		} else if (line.substr(0, 9) === 'Ability: ') {
+			line = line.substr(9);
+			curSet.ability = line;
+		} else if (line.substr(0, 5) === 'EVs: ') {
+			line = line.substr(5);
+			var evLines = line.split('/');
+			curSet.evs = {hp: 0, atk: 0, def: 0, spa: 0, spd: 0, spe: 0};
+			for (var j = 0; j < evLines.length; j++) {
+				var evLine = evLines[j].trim();
+				var spaceIndex = evLine.indexOf(' ');
+				if (spaceIndex === -1) continue;
+				var statid = BattleStatIDs[evLine.substr(spaceIndex + 1)];
+				var statval = parseInt(evLine.substr(0, spaceIndex));
+				if (!statid) continue;
+				curSet.evs[statid] = statval;
+			}
+		} else if (line.substr(0, 5) === 'IVs: ') {
+			line = line.substr(5);
+			var ivLines = line.split(' / ');
+			curSet.ivs = {hp: 31, atk: 31, def: 31, spa: 31, spd: 31, spe: 31};
+			for (var j = 0; j < ivLines.length; j++) {
+				var ivLine = ivLines[j];
+				var spaceIndex = ivLine.indexOf(' ');
+				if (spaceIndex === -1) continue;
+				var statid = BattleStatIDs[ivLine.substr(spaceIndex + 1)];
+				var statval = parseInt(ivLine.substr(0, spaceIndex));
+				if (!statid) continue;
+				curSet.ivs[statid] = statval;
+			}
+		} else if (line.match(/^[A-Za-z]+ (N|n)ature/)) {
+			var natureIndex = line.indexOf(' Nature');
+			if (natureIndex === -1) natureIndex = line.indexOf(' nature');
+			if (natureIndex === -1) continue;
+			line = line.substr(0, natureIndex);
+			curSet.nature = line;
+		} else if (line.substr(0, 1) === '-' || line.substr(0, 1) === '~') {
+			line = line.substr(1);
+			if (line.substr(0, 1) === ' ') line = line.substr(1);
+			if (!curSet.moves) curSet.moves = [];
+			if (line.substr(0, 14) === 'Hidden Power [') {
+				var hptype = line.substr(14, line.length - 15);
+				line = 'Hidden Power ' + hptype;
+				if (!curSet.ivs) {
+					curSet.ivs = {};
+					for (var stat in BattleTypeChart[hptype].HPivs) {
+						curSet.ivs[stat] = BattleTypeChart[hptype].HPivs[stat];
+					}
+				}
+			}
+			if (line === 'Frustration') {
+				curSet.happiness = 0;
+			}
+			curSet.moves.push(line);
+		}
+	}
+	return team;
+};
+
+var packTeam = exports.packTeam = function (team) {
+	var buf = '';
+	if (!team) return '';
+	for (var i = 0; i < team.length; i++) {
+		var set = team[i];
+		if (buf) buf += ']';
+		// name
+		buf += (set.name || set.species);
+		// species
+		var id = toId(set.species || set.name);
+		buf += '|' + (toId(set.name || set.species) === id ? '' : id);
+		// item
+		buf += '|' + toId(set.item);
+		// ability
+		var template = set.species || set.name;
+		try {
+			template = require('./data/pokedex.js').BattlePokedex[toId(template)];
+		} catch (e) {
+			errlog(e.stack);
+			template = null;
+		}
+		if (!template) return '';
+		var abilities = template.abilities;
+		id = toId(set.ability);
+		if (abilities) {
+			if (abilities['0'] && id === toId(abilities['0'])) {
+				buf += '|';
+			} else if (abilities['1'] && id === toId(abilities['1'])) {
+				buf += '|1';
+			} else if (abilities['H'] && id === toId(abilities['H'])) {
+				buf += '|H';
+			} else {
+				buf += '|' + id;
+			}
+		} else {
+			buf += '|' + id;
+		}
+		// moves
+		buf += '|' + set.moves.map(toId).join(',');
+		// nature
+		buf += '|' + set.nature;
+		// evs
+		var evs = '|';
+		if (set.evs) {
+			evs = '|' + (set.evs['hp'] || '') + ',' + (set.evs['atk'] || '') + ',' + (set.evs['def'] || '') + ',' + (set.evs['spa'] || '') + ',' + (set.evs['spd'] || '') + ',' + (set.evs['spe'] || '');
+		}
+		if (evs === '|,,,,,') {
+			buf += '|';
+		} else {
+			buf += evs;
+		}
+		// gender
+		if (set.gender && set.gender !== template.gender) {
+			buf += '|' + set.gender;
+		} else {
+			buf += '|';
+		}
+		// ivs
+		var ivs = '|';
+		if (set.ivs) {
+			ivs = '|' + (set.ivs['hp'] === 31 || set.ivs['hp'] === undefined ? '' : set.ivs['hp']) + ',' + (set.ivs['atk'] === 31 || set.ivs['atk'] === undefined ? '' : set.ivs['atk']) + ',' + (set.ivs['def'] === 31 || set.ivs['def'] === undefined ? '' : set.ivs['def']) + ',' + (set.ivs['spa'] === 31 || set.ivs['spa'] === undefined ? '' : set.ivs['spa']) + ',' + (set.ivs['spd'] === 31 || set.ivs['spd'] === undefined ? '' : set.ivs['spd']) + ',' + (set.ivs['spe'] === 31 || set.ivs['spe'] === undefined ? '' : set.ivs['spe']);
+		}
+		if (ivs === '|,,,,,') {
+			buf += '|';
+		} else {
+			buf += ivs;
+		}
+		// shiny
+		if (set.shiny) {
+			buf += '|S';
+		} else {
+			buf += '|';
+		}
+		// level
+		if (set.level && set.level !== 100) {
+			buf += '|' + set.level;
+		} else {
+			buf += '|';
+		}
+		// happiness
+		if (set.happiness !== undefined && set.happiness !== 255) {
+			buf += '|' + set.happiness;
+		} else {
+			buf += '|';
+		}
+	}
+	return buf;
+};
+
+var fastUnpackTeam = exports.fastUnpackTeam = function (buf) {
+	if (!buf) return null;
+
+	var team = [];
+	var i = 0, j = 0;
+
+	while (true) {
+		var set = {};
+		team.push(set);
+
+		// name
+		j = buf.indexOf('|', i);
+		set.name = buf.substring(i, j);
+		i = j + 1;
+
+		// species
+		j = buf.indexOf('|', i);
+		set.species = buf.substring(i, j) || set.name;
+		i = j + 1;
+
+		// item
+		j = buf.indexOf('|', i);
+		set.item = buf.substring(i, j);
+		i = j + 1;
+
+		// ability
+		j = buf.indexOf('|', i);
+		var ability = buf.substring(i, j);
+		var template = set.species;
+		try {
+			template = require('./data/pokedex.js').BattlePokedex[toId(template)];
+		} catch (e) {
+			errlog(e.stack);
+			template = null;
+		}
+		set.ability = (template.abilities && ability in {'': 1, 0: 1, 1: 1, H: 1} ? template.abilities[ability || '0'] : ability);
+		i = j + 1;
+
+		// moves
+		j = buf.indexOf('|', i);
+		set.moves = buf.substring(i, j).split(',');
+		i = j + 1;
+
+		// nature
+		j = buf.indexOf('|', i);
+		set.nature = buf.substring(i, j);
+		i = j + 1;
+
+		// evs
+		j = buf.indexOf('|', i);
+		if (j !== i) {
+			var evs = buf.substring(i, j).split(',');
+			set.evs = {
+				hp: Number(evs[0]) || 0,
+				atk: Number(evs[1]) || 0,
+				def: Number(evs[2]) || 0,
+				spa: Number(evs[3]) || 0,
+				spd: Number(evs[4]) || 0,
+				spe: Number(evs[5]) || 0
+			};
+		}
+		i = j + 1;
+
+		// gender
+		j = buf.indexOf('|', i);
+		if (i !== j) set.gender = buf.substring(i, j);
+		i = j + 1;
+
+		// ivs
+		j = buf.indexOf('|', i);
+		if (j !== i) {
+			var ivs = buf.substring(i, j).split(',');
+			set.ivs = {
+				hp: ivs[0] === '' ? 31 : Number(ivs[0]),
+				atk: ivs[1] === '' ? 31 : Number(ivs[1]),
+				def: ivs[2] === '' ? 31 : Number(ivs[2]),
+				spa: ivs[3] === '' ? 31 : Number(ivs[3]),
+				spd: ivs[4] === '' ? 31 : Number(ivs[4]),
+				spe: ivs[5] === '' ? 31 : Number(ivs[5])
+			};
+		}
+		i = j + 1;
+
+		// shiny
+		j = buf.indexOf('|', i);
+		if (i !== j) set.shiny = true;
+		i = j + 1;
+
+		// level
+		j = buf.indexOf('|', i);
+		if (i !== j) set.level = parseInt(buf.substring(i, j), 10);
+		i = j + 1;
+
+		// happiness
+		j = buf.indexOf(']', i);
+		if (j < 0) {
+			if (buf.substring(i)) {
+				set.happiness = Number(buf.substring(i));
+			}
+			break;
+		}
+		if (i !== j) set.happiness = Number(buf.substring(i, j));
+		i = j + 1;
+	}
+	return team;
+};
+
+var teamOverview = exports.teamOverview = function (buf) {
+	var team = fastUnpackTeam(buf);
+	if (!team) return '(empty)';
+	var pokes = [];
+	for (var i = 0; i < team.length; i++) {
+		pokes.push(team[i].species);
+	}
+	if (!pokes.length) return '(empty)';
+	return pokes.join(', ');
+};
